LayoutTests/accessibility/element-reflection-ariaactivedescendant-expected.txt

Checks that element reflection is exposed to the a11y tree for 'ariaActiveDescendantElement'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axTarget1.selectedChildrenCount is 1
PASS axTarget1.selectedChildAtIndex(0).isEqual(axDescendant1) is true
PASS axTarget2.selectedChildrenCount is 1
PASS axTarget2.selectedChildAtIndex(0).isEqual(axDescendantFirstChild) is true
PASS axTarget2.selectedChildrenCount is 1
PASS axTarget2.selectedChildAtIndex(0).isEqual(axDescendant2) is true
PASS axTarget3.selectedChildrenCount is 1
PASS axTarget3.selectedChildAtIndex(0).isEqual(axDescendant3) is true
PASS successfullyParsed is true

TEST COMPLETE
